ci: compliance: split LicenseAndCopyrightCheck into compliance/ directory
Move the LicenseAndCopyrightCheck checker out of the monolithic
check_compliance.py and into
scripts/ci/compliance/license_and_copyright.py, using the new
auto-loading infrastructure.

The now-unused imports (reuse.project, reuse.report, Iterable) are
removed from check_compliance.py.
